perception_service/requirements.txt
fastapi==0.78.0
uvicorn==0.17.6
aiokafka==0.7.2
mysql-connector-python==8.0.29
python-dotenv==0.20.0
pydantic==1.9.0
opencv-python==4.5.5.62  # 视觉处理
pyaudio==0.2.11          # 音频处理
SpeechRecognition==3.8.1 # 语音识别
redis==4.1.0
numpy>=1.22,<2.0.0
python-multipart==0.0.5
openai-whisper   
prometheus-client    
pytesseract
